analysis_server/requirements.txt
gunicorn
Django
djangorestframework
requests
django-heroku